requirements.txt
# Core ML and Computer Vision
ultralytics>=8.0.196
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0

# Object Tracking
filterpy>=1.4.5
scipy>=1.10.0

# Label Studio Integration
label-studio>=1.9.0
label-studio-sdk>=0.0.31

# Data Processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Video Processing
moviepy>=1.0.3

# Configuration and Utilities
pyyaml>=6.0
tqdm>=4.65.0
argparse

# API and Web
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0